Add support for wget2
This commit adds support for new rewrite of wget named `wget2`
Some distros like Fedora Silverblue symlink wget2 to `wget`,
but unfortunately it's not parameter compatible.
Running wget2 with wget parameters would fail and nothing would
be downloadable.

To fix this, we check for wget2 and use it directly,
otherwise we assume `wget` is the normal wget v1.x
Additionally, to avoid further problems on different systems,
we now prefer `curl` over `wget`.